fix: actually put the test back in the right place
